+++ Source/SourceExpander.Embedder.Analyzer/Diagnostics/DiagnosticDescriptors.cs
@@ -4,7 +4,7 @@ namespace SourceExpander.Diagnostics
 {
     public static class DiagnosticDescriptors
     {
-        internal static readonly DiagnosticDescriptor EMBEDDER0001_UsingStatic = new DiagnosticDescriptor(
+        internal static readonly DiagnosticDescriptor EMBEDDER0001_UsingStatic = new(
             "EMBEDDER0001",
             new LocalizableResourceString(
                 nameof(DiagnosticsResources.EMBEDDER0001_Title),
@@ -18,7 +18,7 @@ namespace SourceExpander.Diagnostics
             DiagnosticSeverity.Warning,
             isEnabledByDefault: true
             );
-        internal static readonly DiagnosticDescriptor EMBEDDER0002_UsingAlias = new DiagnosticDescriptor(
+        internal static readonly DiagnosticDescriptor EMBEDDER0002_UsingAlias = new(
             "EMBEDDER0002",
             new LocalizableResourceString(
                 nameof(DiagnosticsResources.EMBEDDER0002_Title),
@@ -32,7 +32,7 @@ namespace SourceExpander.Diagnostics
             DiagnosticSeverity.Warning,
             isEnabledByDefault: true
             );
-        internal static readonly DiagnosticDescriptor EMBEDDER0003_NullableProject = new DiagnosticDescriptor(
+        internal static readonly DiagnosticDescriptor EMBEDDER0003_NullableProject = new(
             "EMBEDDER0003",
             new LocalizableResourceString(
                 nameof(DiagnosticsResources.EMBEDDER0003_Title),
@@ -46,7 +46,7 @@ namespace SourceExpander.Diagnostics
             DiagnosticSeverity.Warning,
             isEnabledByDefault: true
             );
-        internal static readonly DiagnosticDescriptor EMBEDDER0004_NullableDirective = new DiagnosticDescriptor(
+        internal static readonly DiagnosticDescriptor EMBEDDER0004_NullableDirective = new(
             "EMBEDDER0004",
             new LocalizableResourceString(
                 nameof(DiagnosticsResources.EMBEDDER0004_Title),
@@ -54,6 +54,21 @@ namespace SourceExpander.Diagnostics
                 typeof(DiagnosticsResources)),
             new LocalizableResourceString(
                 nameof(DiagnosticsResources.EMBEDDER0004_Title),
+                DiagnosticsResources.ResourceManager,
+                typeof(DiagnosticsResources)),
+            "Compilation",
+            DiagnosticSeverity.Warning,
+            isEnabledByDefault: true
+            );
+
+        internal static readonly DiagnosticDescriptor EMBEDDER0005_ExpandEmbedded = new(
+            "EMBEDDER0005",
+            new LocalizableResourceString(
+                nameof(DiagnosticsResources.EMBEDDER0005_Title),
+                DiagnosticsResources.ResourceManager,
+                typeof(DiagnosticsResources)),
+            new LocalizableResourceString(
+                nameof(DiagnosticsResources.EMBEDDER0005_Title),
                 DiagnosticsResources.ResourceManager,
                 typeof(DiagnosticsResources)),
             "Compilation",